+++ node/directives/index.ts
@@ -1,7 +1,9 @@
 import { WithSession } from './withSession'
 import { WithSender } from './withSender'
+import { WithUserPermissions } from './withUserPermissions'
 
 export const schemaDirectives = {
   withSession: WithSession,
   withSender: WithSender,
+  withUserPermissions: WithUserPermissions,
 }