+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Getting `priceTables` from `vtex.b2b-organizations-graphql`
+
 ## [1.4.1] - 2021-09-03
 
 - Ignore errors when using `withUserPermissions`